matplotlib: Check agains PyCXX that breaks builds
Matplotlib will use the CXX python package, if it finds it. And then it
is reported that matplotlib build will fail (I reproduced this, too).

Closes #56

However, this check is only implemented for Python 2.x. Given that
PyCXX is basically unmaintained, I hope it will be very rare to see on
Python 3.x. Or, if it is there and matplotlib is built for both, 2.x
and 3.x, the Requirement will warn about it for the 2.x case, so we
know where to debug.